Conflicts:
	DVA222_Project/DVA222_Project.vcxproj
	DVA222_Project/DVA222_Project.vcxproj.filters

+++ DVA222_Project/Graphix.cpp
@@ -60,7 +60,7 @@ void FillRectangle(int x, int y, int width, int height )
 void DrawString(string text, int x, int y)
 {
     glRasterPos2i(y,x);
-    for(int i = 0; i < text.length(); i++){ 
+    for(int i = 0; i < (int)text.length(); i++){ 
         glutBitmapCharacter(GLUT_BITMAP_HELVETICA_12, text.data()[i]);
     }
 }
